State and useState() The task object map/array we created in last commit, it was separate from our component. We want this to be part of our state So we pass the tasks object to a react hook called "useState()".
Now tasks object is part of our component state

In our return function we use the tasks to access the tasks object.
Since tasks object is immutable we don't use tasks.push() to update or add anything new to tasks object.

We use the "setTasks" we defined when we called "useState()" hook.