Detete task and prop drilling Previosuly we added the red croos icon to our each task. Now we want to actually remove the task from display page when we click ont he red cross
But we don't have access to that event in our Task.js component because the events are stored in the App.js component.

One way to do that is using context api or redux to access the state within components but for this project I'm not using redux becuase it could be bit complicated to learn redux boiler plates and reducer

Instead for now, I will use props. We can send down a function as a prop and fire that off when we click the red cross icon

So we create a delete function and send it down as props to Tasks.js component.

From Tasks.js, we will pass the delete prop down to Task.js component. Now when a cross icon is clicked, it will trigger the delete function

The action of the function will be passed up to App.js which will hide the task we clicked red cross incon on.

So basically props gets passed down to lower components e.g. App.js component to Task.js component and action gets passed up to higher components e.g. Task.js to App.js component.