Optional message if no task
When we delete all task, we want to show a message "Now task to show"

In the App.js if add the logic to the line where we call out Tasks.js compontner i.e. <Tasks />

You will notice that when we reload the page, the tasks comes back it's because it's just UI and the tasks are in the code

We can make it full stack by having a backend e.g. a springboot app api that we can request to using react that will do crud operation on a database or can do it using json server