Form Input State (controlled components)
Each of the input we added in last comming i.e. task name, day&time and reminder check box, each of them need to have their own piece of state

This state is not app level state i.e. not in App.js. This should be component level state i.e. in AddTask.js

Basically as we add input on the input field add task and day&time and check/uncheck reminder checkbox, they will get the current value we put in them

We can see the change with the help of our react chrome extension i.e. react component