useEffect Hook and Fetch tasks from server
The json data we have at http://localhost:5000/tasks. I want to have then displayed on the webpage like it did for the static tasks data we had in App.js

In order for those data form json server to load on our page when our page load we will use a hook called userEffect.