Add task to server
Right now if we add a task, it will add on UI but it's not persisted in the backend i.e. in our jsons server.
If you reload the page the tasks will come back.

Code to do add operation i.e. add task to our json server is in this commit