Toggle reminder on server
Right now if we update a task's reminder, it will add on UI but it's not persisted in the backend i.e. in our jsons server.

If you reload the page it will rever back.

Code to do update reminder i.e. persist the true.false to our json sever is in this commit